Move NoSplit into its own file
