+++ evalml/tests/preprocessing_tests/test_training_validation_split.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from evalml.preprocessing.data_splitters import NoSplit, TrainingValidationSplit
+from evalml.preprocessing.data_splitters import TrainingValidationSplit
 
 
 def test_tvsplit_nsplits():
@@ -68,18 +68,3 @@ def test_tvsplit_always_within_bounds_with_custom_index(random_seed):
     splits = list(splitter.split(X, y=None))
     assert np.all(np.logical_and(splits[0][0] < N, splits[0][0] >= 0))
     assert np.all(np.logical_and(splits[0][1] < N, splits[0][1] >= 0))
-
-
-def test_nosplit_nsplits():
-    assert NoSplit().get_n_splits() == 0
-
-
-def test_nosplit_default():
-    X = pd.DataFrame({"col1": np.arange(0, 10)})
-    y = pd.Series(np.arange(100, 110), name="target")
-    splitter = NoSplit()
-    splits = list(splitter.split(X, y=y))
-    assert len(splits) == 1 and len(splits[0]) == 2
-
-    np.testing.assert_equal(splits[0][0], np.arange(0, 10))
-    np.testing.assert_equal(splits[0][1], [None]*10)
